pyproject.toml
[project]
name = "urban-threadz-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "google-generativeai>=0.8.6",
    "inngest>=0.5.13",
    "llama-index-core>=0.14.12",
    "llama-index-embeddings-google-genai>=0.3.1",
    "llama-index-readers-file>=0.5.6",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]
